Moved only redux-utils function to the file that called it and removed the utils file
